+++ NEWS.md
@@ -1,3 +1,8 @@
+# SkeletalVis 0.1.1
+
+* Added option to update SkeletalVis data
+* Updated linked SkeletalVis data files
+
 # SkeletalVis 0.1.0
 
 * Initial CRAN submission.